refactor gbs-occt add cmake configs for mesh, io, occt and render extensions
