+++ gbs-io/include/gbs/third_parties/magic_enum.hpp
