+++ gbs-mesh/gbs-meshConfig.cmake.in
@@ -1,0 +1,12 @@
+# gbs-mesh cmake module
+# This module sets the following variables in your project:
+#
+#   gbs-mesh_FOUND - true if gbs found on the system
+#   gbs-mesh_INCLUDE_DIRS - the directory containing gbs headers
+
+@PACKAGE_INIT@
+
+if(NOT TARGET gbs::mesh)
+    include("${CMAKE_CURRENT_LIST_DIR}/gbs-meshTargets.cmake")
+    get_target_property(gbs-mesh_INCLUDE_DIRS gbs::mesh INTERFACE_INCLUDE_DIRECTORIES)
+endif()
