+++ gbs-occt/gbs-occtConfig.cmake.in
@@ -1,0 +1,14 @@
+# gbs-occt cmake module
+# This module sets the following variables in your project:
+#
+#   gbs-occt_FOUND - true if gbs found on the system
+#   gbs-occt_INCLUDE_DIRS - the directory containing gbs headers
+
+@PACKAGE_INIT@
+
+if(NOT TARGET gbs::occt)
+    find_package(OpenCASCADE QUIET)
+
+    include("${CMAKE_CURRENT_LIST_DIR}/gbs-occtTargets.cmake")
+    get_target_property(gbs-occt_INCLUDE_DIRS gbs::occt INTERFACE_INCLUDE_DIRECTORIES)
+endif()
