+++ gbs-occt/include/gbs/occt/brepbuilders.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <gbs-occt/containers.h>
+#include <gbs/occt/containers.h>
 #include <BRepBuilderAPI_MakeWire.hxx>
 #include <BRepBuilderAPI_MakeFace.hxx>
 #include <BRepBuilderAPI_MakeEdge.hxx>
