+++ gbs-occt/src/surfacesbuild.cpp
@@ -1,9 +1,9 @@
-#include <gbs-occt/surfacesbuild.h>
-#include <gbs-occt/curvesbuild.h>
+#include <gbs/occt/surfacesbuild.h>
+#include <gbs/occt/curvesbuild.h>
 #include <Geom_RectangularTrimmedSurface.hxx>
 #include <Geom_Plane.hxx>
 #include <GeomFill_NSections.hxx>
-#include <gbs-occt/containers.h>
+#include <gbs/occt/containers.h>
 #include <Geom_BSplineSurface.hxx>
 #include <exception>
