+++ gbs-occt/tests/tests_geomprim.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
-#include <gbs-occt/geomprim.h>
-#include <gbs-occt/containers.h>
+#include <gbs/occt/geomprim.h>
+#include <gbs/occt/containers.h>
 #include <TColgp_Array1OfPnt.hxx>
 #include <TColgp_Array1OfPnt2d.hxx>
 #include <gp_Pnt2d.hxx>
